Proto new package drasil-code-base, still a huge mess
