+++ code/drasil-code-base/README.md
@@ -1,0 +1,16 @@
+--------------------------------------------------
+### Summary of Folder Structure and File Contents
+Last updated: April 17, 2019
+--------------------------------------------------
+
+**Build**
+  - Contains the code for Drasil
+ 
+README.md
+  - This file
+
+drasil-build.cabal
+  - Cabal file, used by stack to build drasil-build
+
+stack.yaml
+  - Used by Stack
