+++ code/drasil-code-base/drasil-code-base.cabal
@@ -1,0 +1,45 @@
+cabal-version: 1.12
+
+-- This file has been generated from package.yaml by hpack version 0.33.0.
+--
+-- see: https://github.com/sol/hpack
+--
+-- hash: 9beb582683ddb168a70f7b3638b66a0f27150bb7bea5644391727f6d77367819
+
+name:           drasil-code-base
+version:        0.1.9.0
+synopsis:       A framework for code and document generation for scientific software - Build System SubPackage
+description:    Please see the README on GitHub at <https://github.com/JacquesCarette/Drasil#readme>
+homepage:       https://jacquescarette.github.io/Drasil/
+bug-reports:    https://github.com/JacquesCarette/Drasil/issues
+author:         Dan Szymczak, Steven Palmer, Jacques Carette, Spencer Smith
+maintainer:     Jacques Carette
+build-type:     Simple
+
+source-repository head
+  type: git
+  location: https://github.com/JacquesCarette/Drasil
+
+library
+  exposed-modules:
+      Language.Drasil.Chunk.CodeBase
+      Language.Drasil.Code.Expr
+      Language.Drasil.Code.Expr.Extract
+      Language.Drasil.Code.Expr.Render
+  other-modules:
+      Paths_drasil_code_base
+  hs-source-dirs:
+      ./
+  ghc-options: -Wall -Wredundant-constraints
+  build-depends:
+      base >=4.7 && <5
+    , containers
+    , drasil-database
+    , drasil-lang
+    , drasil-utils
+    , lens
+    , pretty
+  if false
+    other-modules:
+        Paths_drasil_build
+  default-language: Haskell2010
