+++ code/drasil-code-base/package.yaml
@@ -1,8 +1,8 @@
-name:                drasil-printers
-version:             0.1.10.0
+name:                drasil-code-base
+version:             0.1.9.0
 author:              "Dan Szymczak, Steven Palmer, Jacques Carette, Spencer Smith"
 maintainer:          "Jacques Carette"
-synopsis:	           A framework for code and document generation for scientific software - Printers SubPackage
+synopsis:	           A framework for code and document generation for scientific software - Build System SubPackage
 github:              JacquesCarette/Drasil
 homepage:            https://jacquescarette.github.io/Drasil/
 description:         Please see the README on GitHub at <https://github.com/JacquesCarette/Drasil#readme>
@@ -11,9 +11,9 @@ extra-source-files: []
 
 dependencies:
 - base >= 4.7 && < 5
+- containers
 - lens
 - pretty
-- drasil-code-base
 - drasil-database
 - drasil-lang
 - drasil-utils
@@ -24,8 +24,6 @@ ghc-options:
 
 library:
   source-dirs: ./
-  exposed-modules:
-  - Language.Drasil.Printers
   when:
   - condition: false
-    other-modules: Paths_drasil_printers
+    other-modules: Paths_drasil_build
