+++ code/drasil-code/package.yaml
@@ -18,13 +18,14 @@ dependencies:
 - directory
 - filepath
 - split
-- drasil-lang
-- drasil-utils
 - drasil-build
-- drasil-theory
+- drasil-code-base
 - drasil-database
-- drasil-printers
 - drasil-gool
+- drasil-lang
+- drasil-printers
+- drasil-theory
+- drasil-utils
 
 ghc-options:
 - -Wall
@@ -43,9 +44,6 @@ library:
     - Language.Drasil.Code.Code
     - Language.Drasil.Code.CodeQuantityDicts
     - Language.Drasil.Code.DataDesc
-    - Language.Drasil.Code.Expr
-    - Language.Drasil.Code.Expr.Extract
-    - Language.Drasil.Code.Expr.Render
     - Language.Drasil.Code.ExternalLibrary
     - Language.Drasil.Code.ExternalLibraryCall
     - Language.Drasil.Code.ExtLibImport
@@ -106,9 +104,6 @@ executables:
     - Language.Drasil.Code.Code
     - Language.Drasil.Code.CodeQuantityDicts
     - Language.Drasil.Code.DataDesc
-    - Language.Drasil.Code.Expr
-    - Language.Drasil.Code.Expr.Extract
-    - Language.Drasil.Code.Expr.Render
     - Language.Drasil.Code.ExternalLibrary
     - Language.Drasil.Code.ExternalLibraryCall
     - Language.Drasil.Code.ExtLibImport
