+++ code/drasil-printers/Language/Drasil/Plain/Print.hs
@@ -1,13 +1,14 @@
 module Language.Drasil.Plain.Print (
-  Linearity(..), exprDoc, sentenceDoc, symbolDoc, unitDoc
+  Linearity(..), exprDoc, codeExprDoc, sentenceDoc, symbolDoc, unitDoc
 ) where
 
 import Database.Drasil (ChunkDB)
 import Language.Drasil (Sentence, Special(..), Stage(..), Symbol, USymb(..))
 import qualified Language.Drasil as L (Expr)
+import qualified Language.Drasil.Code.Expr as C (CodeExpr)
 import Language.Drasil.Printing.AST (Expr(..), Spec(..), Ops(..), Fence(..), 
   OverSymb(..), Fonts(..), Spacing(..))
-import Language.Drasil.Printing.Import (expr, spec, symbol)
+import Language.Drasil.Printing.Import (expr, codeExpr, spec, symbol)
 import Language.Drasil.Printing.PrintingInformation (PrintingConfiguration(..), 
   PrintingInformation(..), Notation(Scientific))
 
@@ -29,6 +30,10 @@ plainConfiguration = PC Scientific
 -- | Create expressions for a document in 'Doc' format.
 exprDoc :: ChunkDB -> Stage -> Linearity -> L.Expr -> Doc
 exprDoc db st f e = pExprDoc f (expr e (PI db st plainConfiguration))
+
+-- | Create code expressions for a document in 'Doc' format.
+codeExprDoc :: ChunkDB -> Stage -> Linearity -> C.CodeExpr -> Doc
+codeExprDoc db st f e = pExprDoc f (codeExpr e (PI db st plainConfiguration))
 
 -- | Create sentences for a document in 'Doc' format.
 sentenceDoc :: ChunkDB -> Stage -> Linearity -> Sentence -> Doc
