+++ code/drasil-printers/Language/Drasil/Printers.hs
@@ -11,13 +11,14 @@ module Language.Drasil.Printers (
   , makeMd, introInfo, verInfo, unsupOS, regularSec, instDoc, extLibSec
   -- Plain
   -- Print
-  , Linearity(..), sentenceDoc, exprDoc, symbolDoc, unitDoc
+  , Linearity(..)
+  , sentenceDoc, exprDoc, codeExprDoc, symbolDoc, unitDoc
   -- TeX
   -- Print
   , genTeX
   , PrintingInformation(..)
   , HasPrintingOptions (..)
-  , Notation(Scientific, Engineering)
+  , Notation(..)
   , defaultConfiguration
   )
   where
@@ -25,10 +26,10 @@ module Language.Drasil.Printers (
 import Language.Drasil.Format (Format(TeX, HTML))
 import Language.Drasil.HTML.CSS (makeCSS)
 import Language.Drasil.HTML.Print (genHTML)
-import Language.Drasil.Markdown.CreateMd (makeMd, introInfo, verInfo, unsupOS, extLibSec, 
-  instDoc, regularSec)
-import Language.Drasil.Plain.Print (Linearity(..), sentenceDoc, exprDoc, 
-  symbolDoc, unitDoc)
+import Language.Drasil.Markdown.CreateMd (makeMd, introInfo, verInfo, unsupOS,
+  extLibSec, instDoc, regularSec)
+import Language.Drasil.Plain.Print (Linearity(..), sentenceDoc, exprDoc,
+  codeExprDoc, symbolDoc, unitDoc)
 import Language.Drasil.TeX.Print (genTeX)
 import Language.Drasil.Printing.PrintingInformation (PrintingInformation(..),
-  HasPrintingOptions(..), Notation(Scientific, Engineering), defaultConfiguration)
+  HasPrintingOptions(..), Notation(..), defaultConfiguration)
