+++ code/stack.yaml
@@ -41,6 +41,7 @@ packages:
 - drasil-utils/
 - drasil-gool/
 - drasil-build/
+- drasil-code-base/
 - drasil-code/
 - drasil-theory/
 - drasil-database/
